Enable browser refresh/back/forward
The url wasn't updating at all, even as the page changed to display
results. This udpates the url with each query and also enables browser
actions (back/foward/refresh). In order to do this, it uses session
storage to cache the saved state (keyed to the selected color) so that
back/forward/refresh will work as expected.

However, the cache doesn't work perfectly. The flask server retrieves
images from the API in a random sort, which means that two calls for the
same color will likely generate different results. However, because
session storage uses the color as the key, any duplicate search for a
color will overwrite the first set of results, and will return the same
results for both calls, instead of maintaining the actual history. This
should definitely be fixed at some point.

Signed-off-by: isabeldepapel <isuchanek@gmail.com>