Fix linting
Linting had stopped working because it wasn't compatible with
Typescript. This updates the eslint config to use the wesbos typescript
extension in beta, and then also disables some eslint rules that
conflicted with their (more rigorous) typescript equivalents.

This also fixes all the linting failures that showed up once linting was
working again, including a couple of accessbility issues with the
non-interactive elements.

Signed-off-by: isabeldepapel <isuchanek@gmail.com>